Changes according to PR comments

+++ include/magnetometer_calibration.h
@@ -18,8 +18,6 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <stdint.h>
-
 #pragma once
 
 typedef struct compassBiasEstimator_s {
@@ -31,4 +29,4 @@ typedef struct compassBiasEstimator_s {
 void compassBiasEstimatorInit(compassBiasEstimator_t *compassBiasEstimator, const float lambda_min, const float p0);
 void compassBiasEstimatorUpdate(compassBiasEstimator_t *compassBiasEstimator, const float lambda_min, const float p0);
 void compassBiasEstimatorApply(compassBiasEstimator_t *compassBiasEstimator, float *mag, const float *dmag, const float *gyro);
-void compassBiasEstimatorSolveIterative(compassBiasEstimator_t *compassBiasEstimator, const float *e, float *zn, const float *gyro, const uint8_t k, const uint8_t i, const uint8_t j, const float sign);
+void compassBiasEstimatorSolveIterative(compassBiasEstimator_t *compassBiasEstimator, const float *e, float *zn, const float *gyro, const unsigned k, const unsigned i, const unsigned j, const float sign);
